Add filter to bool variables so ES will not complain
ES fails to properly parse "True" or "False" when they are capitalized. Ansible will insert unquoted values of "true" or "false" into templates as "True" or "False" (without the quotes). With this change, it is no longer necessary to quote these values; they will work properly in the ES template either way.