Improve swap configuration, reformat syntax
Change logic on swap enable/disable tasks from string comparison to
'true' to using a Jinja filter that converts the string to a bool.

Reformat syntax of playbook to be multiline indented. This makes reading
diffs much easier.

Add task to enable swap if es_disable_swap is false.

Add handler to turn swap on.