Add conditional to restart elasticsearh task
es_restart_on_config_change allows control of node restart or config file change.
The main use case for this is updating minimum_master_nodes in the config file
so the cluster will come up if the total number of nodes is reduced. This setting
is transient and can be updated dynamically using the REST API.