Add tasks to main.yml to handle SSL proxy config
Add task to set es_http_proxy_port when es_ssl_proxy is True.

    Get current HTTP port from /etc/elasticsearch/elasticsearch.yml to trigger an
explicit restart if it is different from what is being set in the config file and
tested with a wait_for task. If ES is not restarted when the HTTP port is changed,
then the wait_for task will fail because it will be looking at the new port which is
not active yet.

     Add task to remove elasticsearch nginx config when not proxying. This continuse my theme of making playbooks that cleanup after themselves.

    Instead of calculating the value of es_http_proxy_port in each task, do it once
in a set_fact task and only if using SSL proxy.