Set client_max_body_size in nginx config
The same variable is used for the nginx config as well as the Elasticsearch config,
which means the unit suffix had to be hard coded into the template. While less
than ideal, it fixes the main problem of Logstash trying to shove huge amounts
of logs in a PUT and being denied by nginx.