Stop being fancy with the logic on restarting Elasticsearch
The original intent was to avoid an unecessary reastart on the initial installation
of Elasticsearch, and only restart mid-play if es_http_port changed because the
wait_for task will fail if the port changes and Elasticsearch is not restarted. I was
not able to get the logic of the conditional to line up with the desired result,
which was to avoid an unecessary restart. For the sake of consistency, I have
made the conditional fire if the new es_http_port is different than the current one.